+++ tools/ccgen_impl.cc
@@ -72,8 +72,10 @@ int Main(int argc, const char* const argv[]) {
                 << " [--codec webp2 {444|420} {effort}]" << std::endl
                 << " [--codec jpegxl 444 {effort}]" << std::endl
                 << " [--codec avif {444|420} {effort}]" << std::endl
-                << " [--codec slimavif {444|420} {effort}]" << std::endl
-                << " [--codec slimav2f {444|420} {effort}]" << std::endl
+                << " [--codec avifexp {444|420} {effort}]" << std::endl
+                << "   (mini, YCgCo-Re, AV1)" << std::endl
+                << " [--codec avifavm {444|420} {effort}]" << std::endl
+                << "   (mini, YCgCo-Re, AVM)" << std::endl
                 << " [--codec combination {444|420} {effort}]" << std::endl
                 << " [--codec jpegturbo {444|420}]" << std::endl
                 << " [--codec jpegli {444|420}]" << std::endl
@@ -118,12 +120,12 @@ int Main(int argc, const char* const argv[]) {
           codec_settings.push_back({Codec::kJpegXl, subsampling.value, effort});
         } else if (codec == "avif") {
           codec_settings.push_back({Codec::kAvif, subsampling.value, effort});
-        } else if (codec == "slimavif") {
+        } else if (codec == "avifexp") {
           codec_settings.push_back(
-              {Codec::kSlimAvif, subsampling.value, effort});
-        } else if (codec == "slimav2f") {
+              {Codec::kAvifExp, subsampling.value, effort});
+        } else if (codec == "avifavm") {
           codec_settings.push_back(
-              {Codec::kSlimAvifAvm, subsampling.value, effort});
+              {Codec::kAvifAvm, subsampling.value, effort});
         } else if (codec == "combination") {
           codec_settings.push_back(
               {Codec::kCombination, subsampling.value, effort});